Xorg and Wayland are now both systemd user targets
The WM socket is only defined within the windowing environment, so it
has to be imported into systemd in the WM config file for use by other
units like the desktop bar.

As such we need to make sure other units are resilient to the env var
showing up a little late, and we can delay their start to give the WM a
chance to start.

The graphical session delay provides this timing.