Use sway scale 2 and adjust fonts to compensate
Sway wants to default to a scale 2 on hi-dpi screens. Rather then
fighting it adjust fonts accordingly. When entering tablet mode set
scale to 3 to make the screen usable with fingers.

This eliminates the need for much of the custom tablet-mode
configuration and custom font size configuration for waybar.

It would be a lot nicer if we could tell sway output to reset to its
default scale on tablet-off, but I don't think that's possible atm.