Fix mounting EFI partition in mount script
The EFI partition on tecopa wasn't mounting during recovery operations
using bin/mount-partitions from recovery media. This was happening
because the EFI partition was never managed by ansible, and thus not
included in the list of known partitions.

This change explicitly marks partitions as managed or unmanaged, and
unmanaged partitions should purely be used for read-only purposes